refactor(rust-dump-exporter): god-file 3분할 - ForeignKeyResolver/dump_progress 분리 (CC-099~104)
- src/core/foreign_key_resolver.py 신규: OrphanRecordInfo, ForeignKeyResolver
  이동. _orphan_join_where 헬퍼로 4곳의 FROM/JOIN/WHERE 조각 공유 (CC-104)
- src/exporters/dump_progress.py 신규: TableProgressTracker,
  _format_import_phase_message, emit_core_event 이동. DumpEventCallbacks
  dataclass 도입 + 이벤트타입별 private 핸들러 dispatch table로 재구성 (CC-103)
- rust_dump_exporter.py: _RustDumpClientBase 도입으로 Exporter/Importer
  __init__/_endpoint 중복 제거 (CC-100). private _run_rust_dump만
  DumpEventCallbacks 번들로 축소, 공개 export_full_schema/export_tables/
  import_dump 시그니처는 불변 (CC-101). DEFAULT_DUMP_THREADS 상수화 (CC-102)
- 하위호환 re-export 유지 (ForeignKeyResolver/OrphanRecordInfo/
  TableProgressTracker/emit_core_event 등) - 기존 5개 소비자 파일 무변경
- CLAUDE.md의 ForeignKeyResolver 위치 서술 갱신
