+++ src/exporters/rust_dump_exporter.py
@@ -13,11 +13,14 @@ from src.core.db_core_service import (
     DbEndpoint,
     normalize_db_engine,
 )
+from src.core.foreign_key_resolver import ForeignKeyResolver, OrphanRecordInfo
 from src.core.logger import get_logger
+from src.exporters.dump_progress import DumpEventCallbacks, TableProgressTracker, emit_core_event
 
 logger = get_logger("rust_dump_exporter")
 
 DEFAULT_DUMP_COMPRESSION = "zstd"
+DEFAULT_DUMP_THREADS = 8
 
 
 def _safe_dump_child_dir(dump_dir: str, table_path: str) -> Optional[Path]:
@@ -47,20 +50,6 @@ def _safe_dump_child_file(dump_dir: str, path: Path) -> Optional[Path]:
     except ValueError:
         return None
     return file_path if file_path.is_file() else None
-
-
-def _format_import_phase_message(event: dict) -> Optional[str]:
-    strategy = str(event.get("strategy") or "")
-    performance = str(event.get("performance") or "")
-    message = str(event.get("message") or "")
-    if (
-        strategy in {"temporary_local_infile", "temporary_local_infile_restore", "insert_fallback"}
-        or performance == "fast_path"
-        or "local_infile" in message
-        or "LOAD DATA LOCAL" in message
-    ):
-        return None
-    return str(event.get("message") or event.get("phase") or "Rust DB Core 작업 중...")
 
 
 def _shutdown_owned_facade(facade: DbCoreFacade, owns_facade: bool) -> None:
@@ -127,205 +116,8 @@ cargo build --manifest-path migration_core/Cargo.toml --release
 """
 
 
-@dataclass
-class OrphanRecordInfo:
-    """Foreign-key orphan record summary."""
-
-    table: str
-    column: str
-    referenced_table: str
-    referenced_column: str
-    orphan_count: int
-    sample_values: List[str]
-    query: str
-
-
-class ForeignKeyResolver:
-    """Foreign-key dependency analysis used before partial table dumps."""
-
-    def __init__(self, connector: MySQLConnector):
-        self.connector = connector
-
-    def get_all_dependencies(self, schema: str) -> Dict[str, Set[str]]:
-        query = """
-        SELECT TABLE_NAME, REFERENCED_TABLE_NAME
-        FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE
-        WHERE TABLE_SCHEMA = %s
-          AND REFERENCED_TABLE_NAME IS NOT NULL
-        """
-        deps: Dict[str, Set[str]] = {}
-        for row in self.connector.execute(query, (schema,)):
-            table = row["TABLE_NAME"]
-            ref_table = row["REFERENCED_TABLE_NAME"]
-            if table != ref_table:
-                deps.setdefault(table, set()).add(ref_table)
-        return deps
-
-    def get_fk_details(self, schema: str) -> List[Dict]:
-        query = """
-        SELECT
-            TABLE_NAME,
-            COLUMN_NAME,
-            REFERENCED_TABLE_NAME,
-            REFERENCED_COLUMN_NAME,
-            CONSTRAINT_NAME
-        FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE
-        WHERE TABLE_SCHEMA = %s
-          AND REFERENCED_TABLE_NAME IS NOT NULL
-        ORDER BY TABLE_NAME, COLUMN_NAME
-        """
-        rows = self.connector.execute(query, (schema,))
-        return [
-            {
-                "table": row["TABLE_NAME"],
-                "column": row["COLUMN_NAME"],
-                "referenced_table": row["REFERENCED_TABLE_NAME"],
-                "referenced_column": row["REFERENCED_COLUMN_NAME"],
-                "constraint_name": row["CONSTRAINT_NAME"],
-            }
-            for row in rows
-        ]
-
-    def generate_orphan_query(
-        self,
-        schema: str,
-        table: str,
-        column: str,
-        ref_table: str,
-        ref_column: str,
-    ) -> str:
-        return f"""SELECT c.*
-FROM `{schema}`.`{table}` c
-LEFT JOIN `{schema}`.`{ref_table}` p ON c.`{column}` = p.`{ref_column}`
-WHERE c.`{column}` IS NOT NULL
-  AND p.`{ref_column}` IS NULL"""
-
-    def find_orphan_records(
-        self,
-        schema: str,
-        tables: Optional[List[str]] = None,
-        sample_limit: int = 5,
-        progress_callback: Optional[Callable[[str], None]] = None,
-    ) -> List[OrphanRecordInfo]:
-        fk_details = self.get_fk_details(schema)
-        if tables:
-            table_set = set(tables)
-            fk_details = [fk for fk in fk_details if fk["table"] in table_set]
-
-        results = []
-        for index, fk in enumerate(fk_details, 1):
-            table = fk["table"]
-            column = fk["column"]
-            ref_table = fk["referenced_table"]
-            ref_column = fk["referenced_column"]
-
-            if progress_callback:
-                progress_callback(f"검사 중... ({index}/{len(fk_details)}) {table}.{column}")
-
-            count_query = f"""
-            SELECT COUNT(*) as cnt
-            FROM `{schema}`.`{table}` c
-            LEFT JOIN `{schema}`.`{ref_table}` p ON c.`{column}` = p.`{ref_column}`
-            WHERE c.`{column}` IS NOT NULL
-              AND p.`{ref_column}` IS NULL
-            """
-            count_result = self.connector.execute(count_query)
-            orphan_count = count_result[0]["cnt"] if count_result else 0
-            if orphan_count <= 0:
-                continue
-
-            sample_query = f"""
-            SELECT DISTINCT c.`{column}` as orphan_value
-            FROM `{schema}`.`{table}` c
-            LEFT JOIN `{schema}`.`{ref_table}` p ON c.`{column}` = p.`{ref_column}`
-            WHERE c.`{column}` IS NOT NULL
-              AND p.`{ref_column}` IS NULL
-            LIMIT {sample_limit}
-            """
-            sample_values = [
-                str(row["orphan_value"])
-                for row in self.connector.execute(sample_query)
-            ]
-            results.append(
-                OrphanRecordInfo(
-                    table=table,
-                    column=column,
-                    referenced_table=ref_table,
-                    referenced_column=ref_column,
-                    orphan_count=orphan_count,
-                    sample_values=sample_values,
-                    query=self.generate_orphan_query(schema, table, column, ref_table, ref_column),
-                )
-            )
-        return results
-
-    def export_orphan_report(
-        self,
-        schema: str,
-        output_path: str,
-        tables: Optional[List[str]] = None,
-        progress_callback: Optional[Callable[[str], None]] = None,
-    ) -> Tuple[bool, str, int]:
-        try:
-            orphans = self.find_orphan_records(schema, tables, progress_callback=progress_callback)
-            with open(output_path, "w", encoding="utf-8") as file:
-                file.write("# 고아 레코드 분석 보고서\n")
-                file.write(f"# 스키마: {schema}\n")
-                file.write(f"# 생성일시: {datetime.now().isoformat()}\n")
-                file.write(f"# 발견된 고아 관계: {len(orphans)}건\n")
-                file.write("=" * 80 + "\n\n")
-                if not orphans:
-                    file.write("고아 레코드가 발견되지 않았습니다.\n")
-                else:
-                    total_orphans = sum(item.orphan_count for item in orphans)
-                    file.write(f"총 {total_orphans:,}개의 고아 레코드 발견\n\n")
-                    for index, item in enumerate(orphans, 1):
-                        file.write(
-                            f"## [{index}] {item.table}.{item.column} -> "
-                            f"{item.referenced_table}.{item.referenced_column}\n"
-                        )
-                        file.write(f"   고아 레코드 수: {item.orphan_count:,}건\n")
-                        file.write(f"   샘플 값: {', '.join(item.sample_values)}\n")
-                        file.write("\n   조회 쿼리:\n")
-                        file.write("   ```sql\n")
-                        for line in item.query.split("\n"):
-                            file.write(f"   {line}\n")
-                        file.write("   ```\n\n")
-                        file.write("-" * 80 + "\n\n")
-            return True, f"보고서 저장 완료: {output_path}", len(orphans)
-        except Exception as exc:
-            return False, f"보고서 저장 실패: {exc}", 0
-
-    def get_all_orphan_queries(self, schema: str, tables: Optional[List[str]] = None) -> str:
-        fk_details = self.get_fk_details(schema)
-        if tables:
-            table_set = set(tables)
-            fk_details = [fk for fk in fk_details if fk["table"] in table_set]
-
-        queries = [
-            f"-- 고아 레코드 조회 쿼리 (스키마: {schema})",
-            f"-- 생성일시: {datetime.now().isoformat()}",
-            f"-- FK 관계 수: {len(fk_details)}개",
-            "",
-        ]
-        for index, fk in enumerate(fk_details, 1):
-            table = fk["table"]
-            column = fk["column"]
-            ref_table = fk["referenced_table"]
-            ref_column = fk["referenced_column"]
-            queries.append(f"-- [{index}] {table}.{column} -> {ref_table}.{ref_column}")
-            queries.append(
-                f"""SELECT '{table}.{column}' AS fk_relation, COUNT(*) AS orphan_count
-FROM `{schema}`.`{table}` c
-LEFT JOIN `{schema}`.`{ref_table}` p ON c.`{column}` = p.`{ref_column}`
-WHERE c.`{column}` IS NOT NULL AND p.`{ref_column}` IS NULL;
-"""
-            )
-        return "\n".join(queries)
-
-
-class RustDumpExporter:
-    """Rust DB Core backed dump exporter."""
+class _RustDumpClientBase:
+    """Shared facade/endpoint plumbing for Rust dump exporter/importer."""
 
     def __init__(self, config: RustDumpConfig, facade: Optional[DbCoreFacade] = None):
         self.config = config
@@ -341,6 +133,10 @@ class RustDumpExporter:
             password=self.config.password,
             database=schema,
         )
+
+
+class RustDumpExporter(_RustDumpClientBase):
+    """Rust DB Core backed dump exporter."""
 
     def _resolve_required_tables_from_rust_schema(
         self,
@@ -403,15 +199,12 @@ class RustDumpExporter:
         schema: str,
         output_dir: str,
         tables: Optional[List[str]],
-        threads: int = 8,
+        threads: int = DEFAULT_DUMP_THREADS,
         chunk_size: int = 50000,
         compression: str = DEFAULT_DUMP_COMPRESSION,
-        progress_callback: Optional[Callable[[str], None]] = None,
-        table_progress_callback: Optional[Callable[[int, int, str], None]] = None,
-        detail_callback: Optional[Callable[[dict], None]] = None,
-        table_status_callback: Optional[Callable[[str, str, str], None]] = None,
-        raw_output_callback: Optional[Callable[[str], None]] = None,
+        callbacks: Optional[DumpEventCallbacks] = None,
     ) -> Tuple[bool, str]:
+        callbacks = callbacks or DumpEventCallbacks()
         payload = {
             "source": self._endpoint(schema).to_payload(),
             "output_dir": output_dir,
@@ -424,18 +217,18 @@ class RustDumpExporter:
         if tables:
             payload["tables"] = tables
 
-        if progress_callback:
-            progress_callback(f"Rust DB Core export 시작: {self.config.get_masked_uri()}/{schema}")
+        if callbacks.progress:
+            callbacks.progress(f"Rust DB Core export 시작: {self.config.get_masked_uri()}/{schema}")
 
         result = self.facade.run_dump(
             payload,
             on_event=lambda event: self._emit_core_event(
                 event,
-                progress_callback,
-                table_progress_callback,
-                detail_callback,
-                table_status_callback,
-                raw_output_callback,
+                callbacks.progress,
+                callbacks.table_progress,
+                callbacks.detail,
+                callbacks.table_status,
+                callbacks.raw_output,
             ),
         )
         rows = int(result.get("rows_dumped") or 0)
@@ -450,7 +243,7 @@ class RustDumpExporter:
         self,
         schema: str,
         output_dir: str,
-        threads: int = 8,
+        threads: int = DEFAULT_DUMP_THREADS,
         compression: str = DEFAULT_DUMP_COMPRESSION,
         progress_callback: Optional[Callable[[str], None]] = None,
         table_progress_callback: Optional[Callable[[int, int, str], None]] = None,
@@ -465,11 +258,13 @@ class RustDumpExporter:
                 tables=None,
                 threads=threads,
                 compression=compression,
-                progress_callback=progress_callback,
-                table_progress_callback=table_progress_callback,
-                detail_callback=detail_callback,
-                table_status_callback=table_status_callback,
-                raw_output_callback=raw_output_callback,
+                callbacks=DumpEventCallbacks(
+                    progress=progress_callback,
+                    table_progress=table_progress_callback,
+                    detail=detail_callback,
+                    table_status=table_status_callback,
+                    raw_output=raw_output_callback,
+                ),
             )
             if success:
                 self._write_metadata(output_dir, schema, "full", None)
@@ -486,7 +281,7 @@ class RustDumpExporter:
         schema: str,
         tables: List[str],
         output_dir: str,
-        threads: int = 8,
+        threads: int = DEFAULT_DUMP_THREADS,
         compression: str = DEFAULT_DUMP_COMPRESSION,
         include_fk_parents: bool = True,
         progress_callback: Optional[Callable[[str], None]] = None,
@@ -512,11 +307,13 @@ class RustDumpExporter:
                 tables=final_tables,
                 threads=threads,
                 compression=compression,
-                progress_callback=progress_callback,
-                table_progress_callback=table_progress_callback,
-                detail_callback=detail_callback,
-                table_status_callback=table_status_callback,
-                raw_output_callback=raw_output_callback,
+                callbacks=DumpEventCallbacks(
+                    progress=progress_callback,
+                    table_progress=table_progress_callback,
+                    detail=detail_callback,
+                    table_status=table_status_callback,
+                    raw_output=raw_output_callback,
+                ),
             )
             if success:
                 self._write_metadata(output_dir, schema, "partial", final_tables, added_tables)
@@ -565,23 +362,8 @@ def _mark_non_done_import_results_error(
             table_status_callback(table, "error", message)
 
 
-class RustDumpImporter:
+class RustDumpImporter(_RustDumpClientBase):
     """Rust DB Core backed dump importer."""
-
-    def __init__(self, config: RustDumpConfig, facade: Optional[DbCoreFacade] = None):
-        self.config = config
-        self.facade = facade if facade is not None else DbCoreFacade()
-        self._owns_facade = facade is None
-
-    def _endpoint(self, schema: str) -> DbEndpoint:
-        return DbEndpoint(
-            engine=self.config.engine,
-            host=self.config.host,
-            port=int(self.config.port),
-            user=self.config.user,
-            password=self.config.password,
-            database=schema,
-        )
 
     def _analyze_dump_metadata(self, dump_dir: str) -> Optional[Dict]:
         manifest_path = Path(dump_dir) / "_tunnelforge_dump.json"
@@ -630,7 +412,7 @@ class RustDumpImporter:
         self,
         input_dir: str,
         target_schema: Optional[str] = None,
-        threads: int = 8,
+        threads: int = DEFAULT_DUMP_THREADS,
         import_mode: str = "replace",
         timezone_sql: Optional[str] = None,
         progress_callback: Optional[Callable[[str], None]] = None,
@@ -725,131 +507,6 @@ class RustDumpImporter:
             _shutdown_owned_facade(self.facade, self._owns_facade)
 
 
-class TableProgressTracker:
-    """Table progress metadata helper for the import UI."""
-
-    def __init__(self, metadata: Optional[Dict]):
-        self.chunk_counts = metadata.get("chunk_counts", {}) if metadata else {}
-        self.table_sizes = metadata.get("table_sizes", {}) if metadata else {}
-        self.total_bytes = metadata.get("total_bytes", 0) if metadata else 0
-        self.completed_tables: Set[str] = set()
-
-    def estimate_loading_tables(
-        self,
-        loaded_bytes: int,
-        completed_tables: List[str],
-    ) -> List[Tuple[str, int, int]]:
-        self.completed_tables = set(completed_tables)
-        candidates = [
-            (table, self.table_sizes.get(table, 0), self.chunk_counts.get(table, 1))
-            for table in self.table_sizes
-            if table not in self.completed_tables and self.table_sizes.get(table, 0) > 10_000_000
-        ]
-        candidates.sort(key=lambda item: -item[1])
-        return candidates[:4]
-
-    def get_table_info(self, table_name: str) -> Tuple[int, int]:
-        return self.table_sizes.get(table_name, 0), self.chunk_counts.get(table_name, 1)
-
-    def format_size(self, size_bytes: int) -> str:
-        if size_bytes < 1024:
-            return f"{size_bytes} B"
-        if size_bytes < 1024 * 1024:
-            return f"{size_bytes / 1024:.1f} KB"
-        if size_bytes < 1024 * 1024 * 1024:
-            return f"{size_bytes / (1024 * 1024):.1f} MB"
-        return f"{size_bytes / (1024 * 1024 * 1024):.2f} GB"
-
-
-def emit_core_event(
-    event: Dict,
-    progress_callback: Optional[Callable[[str], None]] = None,
-    table_progress_callback: Optional[Callable[[int, int, str], None]] = None,
-    detail_callback: Optional[Callable[[dict], None]] = None,
-    table_status_callback: Optional[Callable[[str, str, str], None]] = None,
-    raw_output_callback: Optional[Callable[[str], None]] = None,
-    import_results: Optional[dict] = None,
-    table_chunk_progress_callback: Optional[Callable[[str, int, int], None]] = None,
-) -> None:
-    event_type = event.get("event")
-    table = str(event.get("table") or "")
-    if raw_output_callback:
-        raw_output_callback(json.dumps(event, ensure_ascii=False))
-
-    if event_type == "dump_plan":
-        if detail_callback:
-            detail_callback({
-                "event": "dump_plan",
-                "tables_total": int(event.get("tables_total") or 0),
-                "rows_total": int(event.get("rows_total") or 0),
-                "tables": event.get("tables") if isinstance(event.get("tables"), list) else [],
-            })
-    elif event_type == "dump_schedule":
-        if detail_callback:
-            detail_callback({
-                "event": "dump_schedule",
-                "threads": int(event.get("threads") or 0),
-                "table_workers": int(event.get("table_workers") or 0),
-                "range_workers_per_table": int(event.get("range_workers_per_table") or 0),
-                "chunk_size": int(event.get("chunk_size") or 0),
-                "data_format": str(event.get("data_format") or ""),
-                "compression": str(event.get("compression") or ""),
-                "scheduled_tables": event.get("scheduled_tables") if isinstance(event.get("scheduled_tables"), list) else [],
-            })
-    elif event_type == "phase" and progress_callback:
-        message = _format_import_phase_message(event)
-        if message:
-            progress_callback(message)
-    elif event_type == "table_progress":
-        current = int(event.get("current") or 0)
-        total = int(event.get("total") or 0)
-        status = str(event.get("status") or "")
-        ui_status = "loading" if status in ("dumping", "importing") else "done" if status == "completed" else status
-        if table_progress_callback and status == "completed":
-            table_progress_callback(current, total, table)
-        if table_status_callback and table:
-            table_status_callback(table, ui_status, "")
-        if import_results is not None and table:
-            import_results[table] = {"status": ui_status or "loading", "message": ""}
-    elif event_type == "row_progress":
-        rows = int(event.get("table_rows_done") or event.get("rows") or 0)
-        total = int(event.get("table_rows_total") or event.get("total") or 0)
-        overall_rows = int(event.get("overall_rows_done") or 0)
-        overall_total = int(event.get("overall_rows_total") or 0)
-        chunk_rows = int(event.get("chunk_rows") or 0)
-        elapsed_ms = int(event.get("stream_ms") or event.get("read_ms") or event.get("load_ms") or 0)
-        rows_sec = int((chunk_rows * 1000) / elapsed_ms) if chunk_rows and elapsed_ms else 0
-        if overall_total:
-            percent = int((overall_rows / overall_total) * 100)
-        else:
-            percent = int((rows / total) * 100) if total else 0
-        if detail_callback:
-            detail_callback({
-                "event": "row_progress",
-                "table": table,
-                "percent": min(percent, 100),
-                "rows_done": rows,
-                "rows_total": total,
-                "overall_rows_done": overall_rows,
-                "overall_rows_total": overall_total,
-                "chunk_rows": chunk_rows,
-                "rows_sec": rows_sec,
-                "speed": f"{rows_sec:,} rows/s" if rows_sec else "Rust DB Core",
-                "chunk_index": event.get("chunk_index"),
-                "chunks_done": event.get("chunks_done"),
-                "chunks_total": event.get("chunks_total"),
-                "strategy": event.get("strategy"),
-                "stream_ms": event.get("stream_ms"),
-                "read_ms": event.get("read_ms"),
-                "write_ms": event.get("write_ms"),
-                "load_ms": event.get("load_ms"),
-            })
-        chunks_done = int(event.get("chunks_done") or 0)
-        chunks_total = int(event.get("chunks_total") or 0)
-        if table_chunk_progress_callback and table and chunks_done and chunks_total:
-            table_chunk_progress_callback(table, chunks_done, chunks_total)
-
-
 def check_rust_dump() -> Tuple[bool, str]:
     installed, message, _ = RustDumpChecker.check_installation()
     return installed, message
@@ -862,7 +519,7 @@ def export_schema(
     password: str,
     schema: str,
     output_dir: str,
-    threads: int = 8,
+    threads: int = DEFAULT_DUMP_THREADS,
     progress_callback: Optional[Callable[[str], None]] = None,
     engine: str = "mysql",
 ) -> Tuple[bool, str]:
@@ -879,7 +536,7 @@ def export_tables(
     schema: str,
     tables: List[str],
     output_dir: str,
-    threads: int = 8,
+    threads: int = DEFAULT_DUMP_THREADS,
     include_fk_parents: bool = True,
     progress_callback: Optional[Callable[[str], None]] = None,
     engine: str = "mysql",
@@ -903,7 +560,7 @@ def import_dump(
     password: str,
     input_dir: str,
     target_schema: Optional[str] = None,
-    threads: int = 8,
+    threads: int = DEFAULT_DUMP_THREADS,
     import_mode: str = "replace",
     progress_callback: Optional[Callable[[str], None]] = None,
     table_chunk_progress_callback: Optional[Callable[[str, int, int], None]] = None,
